Avoid last minute charge before peak time due to solar shortfall.
Part of the problem is just before 4pm (peak time) we need to hit our charge target. If we just miss it due to solar or usage forecasts being wrong
we end up charging at the last minute. This is also the point in the way when we have the least visibility of future tariff data,
so we don't have flexiblity to use cheaper slots later. By moving the target time back we can have much more flexibilty to adapt to changes